requirements.txt
pdfplumber==0.7.6
pytesseract==0.3.10
Pillow==10.0.1
reportlab==4.0
customtkinter==6.3
openpyxl